everything works as it should (question inside)
but I have this arraylist in the Pause class called backupHS that gives
me an IndexOutOfBoundsException. i think it’s because I’m not
initializing it properly.
but if I initialize it as, say, “backupHS = highscores,” then there are
major display issues (it prints two sets of highscores on top of each
other, which is bad)
does anyone know how to manually add a few values to an arraylist? like
for arrays you can just do {“a”,”b”,c”} to prevent this. i tried double
bracket initialization for the arraylist to no avail…